LPS-160924 moved SelectWithOption Component to Select directory
