+++ modules/apps/object/object-js-components-web/types/src/main/resources/META-INF/resources/components/Select/SelectWithOption.d.ts
@@ -13,7 +13,7 @@
  */
 
 import React from 'react';
-import './Select/index.scss';
+import './index.scss';
 interface SelectWithOptionProps
 	extends React.SelectHTMLAttributes<HTMLSelectElement> {
 	ariaLabel?: string;
